x86: use stdout if insn decoder test is successful
If the instruction decoder test ran successful it prints a message like
this to stderr:
    Succeed: decoded and checked 1767380 instructions

But, as described in "console mode programming user interface guidelines
version 101" which doesn't exist, programs should use stderr for errors
or warnings. We're told about a successful run here, so the instruction
decoder test should use stdout.

Let's fix the typo too, while we're at it.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>