pktcdvd: remove kconfig macro from exported header
pktcdvd.h is exported to userspace. It contains:
    #if defined(CONFIG_CDROM_PKTCDVD_WCACHE)
    #define USE_WCACHING		1
    #else
    #define USE_WCACHING		0
    #endif

Since CONFIG_CDROM_PKTCDVD_WCACHE is not valid in userspace the preprocessor
check wil always evaluate to false when exported. So USE_WCACHING will be
defined to 0 in userspace.

It turns out the kernel can do without USE_WCACHING and use
CONFIG_CDROM_PKTCDVD_WCACHE directly. And that allows us to remove the
preprocessor check and the unreachable #if branch.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>