unistd.h: help unifdef to do the right thing
This header exports this preprocessor check:
    #ifdef CONFIG_MMU

Since CONFIG_MMU is not valid in userspace that check will always evaluate to
false when exported. We should convert this to:
    #if defined(CONFIG_MMU) && defined(__KERNEL__)

That allows unifdef to remove the preprocessor check and the unreachable #if
branch. That way we'll no longer leak this kconfig macro to userspace.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>