x86/auxvec.h: help unifdef to do the right thing
This header exports this preprocessor check:
    #if defined(CONFIG_IA32_EMULATION) || !defined(CONFIG_X86_64)

Since CONFIG_X86_64 is not valid in userspace that check will always evaluate
to true when exported. We should convert this to:
    #if defined(CONFIG_IA32_EMULATION) || !defined(CONFIG_X86_64) || !defined(__KERNEL__)

That allows unifdef to remove the preprocessor check and the unreachable #else
branch. That way we'll no longer leak these two kconfig macros to userspace.

Signed-off-by: Paul Bolle <pebolle@tiscali.nl>